fingerprint: Send dummy notification on unrecognized fingerprint
In authenticate mode only successfully recognized fingerprints are
reported. This makes it impossible to provide feedback for user about
situations where fingerprint acquisition is working as expected, but
unknown finger is pressed on the sensor.

Modify the behavior so that it matches what has been observed in other
devices: Unrecognized fingers are reported via sending authentication
notifications where gid=as-expected and fid=zero.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>