Use SOMC_PLATFORM to guard platform specific sources
TARGET_BOOTLOADER_BOARD_NAME is an aosp build system variable that is
typically used to store a device-specific boardname (e.g. bullhead,
marlin, sailfish), but it was utilized in the sony device projects to
store the platform name instead (e.g. yoshino).  Unfortunately, that
causes updatepackage zips to break, so it seems better to use the sony
variable SOMC_PLATFORM here instead.