loire/tone: Fix auth_id size and location.
The size of auth_id is 64bits, and it's least-significant 32-bits were
stored in the incorrect `result` field.

This change makes auth_id (as returned by getAuthenticatorId) equal to
the value of hat->authenticator_id, allowing applications that use HATs
such as banking apps to work with fingerprint authentication.

TODO: The same should be checked on Yoshino and Nile (with FPC sensors),
and potentially patched in a similar way. It might be spanned over
result and/or unk1 in their fpc_get_db_id_cmd_t struct.

Change-Id: I3766202efd276f7b9467ebb1c29fa195cc91b0d4
Signed-off-by: MarijnS95 <marijns95@gmail.com>