egistec: Add egisap32 TZ command wrapper and datastructures.
Define the initial datastructures used in communication to the TrustZone
app, and some helper functions that fill this buffer for specific
commands.
A locked buffer protects accidental concurrent access to the same ION
buffer, while at the same time handling the offsets of the
datastructures in request and response contexts.

Change-Id: Ic799393cf5926dab9034cdb12184a2cb773b4d25
Signed-off-by: MarijnS95 <marijns95@gmail.com>