egistec: Set userdata-path and store gid.
The path is needed for the TZ-app to know where to store the database in
the right user context. The GID is "ignored" in that the HAL will not
support authentication for multiple users at the same time. In order to
do that, the service needs to call setActiveGroup again with the right
database path for that user.

Change-Id: Ia77f898db45bbd03fe55094328e0060d176ec850
Signed-off-by: MarijnS95 <marijns95@gmail.com>