egistec: Async thread setup with eventfd for lock-free communication.
All long-running operations should happen separately from the thread the
hidl calls are handled on, because the service expects these to return.
Further communication is done through callbacks.
IO multiplexing is used to block on both an event socket and the
fingerprint device, to prevent continuous wakeups to check one or the
other.

Change-Id: I27e6e3ab66f8bd1b71b9ffd409b0b96992eeed87
Signed-off-by: MarijnS95 <marijns95@gmail.com>