Signal onEnumerate end when no fingerprints are available.
Since android Pie this function is actually being called to sync the
prints known by the system (settings_fingerprint.xml) with the ones
existing in the db managed by the FP HAL. On a fresh device with no
fingerprints, the service waits for a few seconds before sending a
cancel event, which in some cases interferes with another running event,
corrupting state. Mitigate this issue by ending the iteration
(remaining=0) with an ignored fid of 0.

Change-Id: I32efa89c124483f3e8c6562e119db86425b169dd
Signed-off-by: MarijnS95 <marijns95@gmail.com>