egistec: Important: Pass locked buffer in Enroll to prevent deadlock.
This is the perfect example of bad API design. Helper functions that
lock the buffer on their own behalf are dangerous.
A safer approach would be to use recursive mutexes, always require a
locked buffer to be passed, or refactor the system entirely to prevent
such things from happening in the first place (buffer locking could be
replaced with more global locking, immediately protecting other shared
variables at the same time).

Change-Id: Ic8516dccae9d686572975a957bac826ca5eb6cda
Signed-off-by: MarijnS95 <marijns95@gmail.com>