get_print_index: Set initial length to MAX, return actual length to caller.
Without the initial length, the tzapp assumes no space is available to
return fids in, returning an error code instead.
After the TZ call, the array is filled and length is set to the number
of fids returned by the device. This field needs to be forwarded to the
returned fpc_fingerprint_index_t struct, otherwise the caller will not
enumerate any prints.

Change-Id: I5483010fc11f96d85de2ebe4eb7a3b2ff80538d8
Signed-off-by: MarijnS95 <marijns95@gmail.com>