egistec: WIP: *Always* cancel when exiting an auth/enroll loop.
If not, the rest of the state breaks in all ways imaginable. Consecutive
calls for example (on completely different/unexpected functions such as
SendInitAuthenticate) suddenly want to wait for a finger (step 4).

WIP Because this is not the most elegant solution. How about returning
an error code and handling Notify+Cancel in the caller? This can then be
unified with timout handling and CheckAndHandleCancel.

Change-Id: Ie23e5b0eadb6968568e9d4bcee5265d26520969a
Signed-off-by: MarijnS95 <marijns95@gmail.com>