egistec: Switch to ET51X ioctls and en-/disable device on another scope.
FPC Hardware is completely powered down when it is not in use, saving
power. Similarly, the IRQ does not need to be enabled manually, but is
enabled as soon as the kernel driver is expected to wait on the hardware
(eg. as soon as calling poll()). This change makes the egistec HAL
compatible with that workflow, as well as switching to the right ioctls.

Change-Id: I2294065fa5fb8a5df65dea1d6da7aef676f4c99f
Signed-off-by: MarijnS95 <marijns95@gmail.com>