fpc auth/enroll: Change power state outside of the loop.
Toggling this within the loop causes a lot of unnecessary changes,
since the device will be turned on immediately after being turned off
when auth was not successful (eg. bad image or waiting for finger off).
Furthermore, this has unexpected side-effects such as changing the IRQ
value because the device leaves the FPC_WAIT_FINGER_ON state. This is a
necessity for future improvements to the poll mechanisn.

Change-Id: Ifb1215322fb975fa5a83d1890805473dda748785
Signed-off-by: MarijnS95 <marijns95@gmail.com>