Extend wakelock through ioctl when waiting for finger_lost.
If an object touches the sensor too long, the wakelock times out,
freezing up the HAL before it reaches the poll. The IRQ trigger from the
sensor hardware is only enabled to wake up the system and HAL from that
point. This unresponsive fingerprint authentication is mitigated by
asking the driver to extend the wakelock timeout.

Change-Id: Iaa6f4bb12596098ca8d2f076ded57b5f421bd48b
Signed-off-by: MarijnS95 <marijns95@gmail.com>