Pass acquire status as normal status intead of vendor code.
This shows states such as FINGERPRINT_ACQUIRED_PARTIAL to the user in
the UI. These states were previously passed as vendor code, which are
ignored when the acquired_info code is not set to ACQUIRED_VENDOR.

Vendor codes should not be passed through anyway because these require
custom strings defined in an overlay.
In fact, the used VENDOR_BASE from libhardware starts at 1000 whereas
the vendor base from the new hidl interface starts at 6.

Change-Id: I0fa704f85202ffb34b913f01e87e0fc4e94ea39f
Signed-off-by: MarijnS95 <marijns95@gmail.com>