loire_tone: Ignore GET_FINGER_STATUS for improved detection.
This command usually returns 0 when the finger is slid on top of the
sensor or placed partially, as opposed to a tap with full coverage
resulting in 1. Even when the finger is not placed optimally,
fpc_capture_image is still able to determine a match or return an
error in case the capture is really insufficient.

Removing this call makes the detection a lot more responsive (and might
explain why it is not existing on the other platforms). In fact, any
invalid response from GET_FINGER_STATUS causes the HAL to wait for the
finger to disappear before trying again, making it even more
inconvenient and slow.

Change-Id: I87d996d65fe491da9d929d47cf35e7cda4e0fbb0