all: Normalize wait_finger_down and capture_image.
Thanks to changes in the detection design for loire_tone, the
imlementation between these and yoshino_nile is now more or less
identical. Getting them in-line with each other prepares for an
eventual merge of the two implementations. Furthermore, this makes
understanding without remembering the differences easier.

Unnecessary commented loops have been removed.

Change-Id: I314520d9ca524619d9861ff04757739e1c3bb298
Signed-off-by: MarijnS95 <marijns95@gmail.com>