tama: Continue capturing when return code equals 3.
This code is returned often on the Tama platform, contrary to other
devices where the capture seems to block until it has a proper image or
can say with confidence that it wasn't able to gather any useful data.

The stock HAL deals with this much the same way: try 10 times with a
result code of 3 before giving up. This is required to prevent looping
(and keeping the CPU awake) indefinitely in case of accidental contact.
It'll restart when the finger/object is removed and placed on the sensor
again.

Change-Id: Iec3fb35a6d3c1f1e5e25c5436bbc2df4273d2204
Signed-off-by: MarijnS95 <marijns95@gmail.com>