egistec: Declare constexpr variables inline to provide storage location.
As usual, static member variables have to be declared *and* defined, no
such exception for constexpr. Instead of defininig these explicitly in a
compilation unit (eg. EGISAPTrustlet.cpp), switch to C++ 17 and use the
inline keyword.

This fixes an incompatibility with android Oreo - apparently the Pie
clang compiler accepts it (as part of a constexpr optimization most
likely).

Change-Id: I329addca5410f6789b7ff2176100f1342ef65635
Signed-off-by: MarijnS95 <marijns95@gmail.com>