Egistec: Correct | EPOLLIN to & EPOLLIN for correctness.
This operation is supposed to be an AND operation, to mask the EPOLLIN
bit out and only proceed if it's set.
In this case it doesn't matter and has no effect on the code flow, since
the poll is only requested for EPOLLIN (meaning that if it triggers, it
can only be that). Fix it nevertheless for code readability and
correctness, as well as to prevent accepting EPOLLERR or EPOLLHUP as
valid triggers.

Change-Id: I6459c35d8c11ad635801816b6588f08b3c10ea8b
Signed-off-by: MarijnS95 <marijns95@gmail.com>