BiometricsFingerprint: Let the service restart the HAL on fatal errors.
Previously, returning a negative status code resulted in the loop
silently breaking, without restarting the operation or informing the
user.

For now, we assume all negative values cannot be handled properly, and
require a full restart (expected cases such as cancellation are dealt
with). This is done by the service when raising ERROR_HW_UNAVAILABLE.
The service tries to initialize the HAL every once in a while: a few
times after the error is raised, as well as when the screen is turned
on. If it fails in the end, the user is properly informed by showing
"Fingerprint hardware not available" and removing the respective icon.

Change-Id: If422d4607bea7b3d408a957b619241b2dcc79541
Signed-off-by: MarijnS95 <marijns95@gmail.com>