BiometricsFingerprint: Transition to idle state before raising events.
The Service using this HAL has the ability to process the next pending
authentication request as soon as an onError or onAuthenticated callback
for the current session is received. The auth/enroll functions have not
returned to their thread loop yet (because they are waiting for the HIDL
callback to finish), with running_state still hanging at STATE_AUTH
instead of STATE_IDLE, resulting in an invalid state transition and
reset of the HAL:
E AOSP FPC HAL (Binder): setState : Invalid state transition to 2 when still processing 2

In order to counter this, set running_state to idle *before* invoking
the callback and returning to the thread loop.

Change-Id: I9c2074fd79b8fa2b421e12495a96d0f8b0fc39bb
Signed-off-by: MarijnS95 <marijns95@gmail.com>