BiometricsFingerprint: Fully abort authentication when auth_step fails.
The partial restart might have been working for Yoshino, but definitely
breaks on Tama. The authentication sequence has to be restarted, or the
TZ will not know what to do with subsequent wait_finger_lost calls and
everything that follows.

Because the change was written for Yoshino specifically - without ever
confirming these errors occuring on other platforms - hide it behind an
ifdef. Raising HW_UNAVAILABLE is more vigorous anyway.

Change-Id: I460fce9e4c66ddb5f940643e4938ebe98762d6fa
Signed-off-by: MarijnS95 <marijns95@gmail.com>