BiometricsFingerprint: Simplify this/device passing once more.
One of the reasons OO exists is to not have to pass a pointer to the
object around, that happens implicitly through `this`.
This patch makes more functions non-static because they need access to
the current state (pretty much everything), while at the same time
removing sdev needlessly being passed around.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>