Egistec: Ganges: Implement ION buffer allocation and handling.
Unfortunately this is necessary on Ganges, while it's not on Nile.

If you follow the documentation for QSEECom_send_modified_cmd carefully,
it seems as if these buffers are supposed to be overlaid at
`cmd_buf_offset` on the buffer passed in or out. This would've made a
lot of extra ION handling obsolete, but it doesn't work like that.
TZapps really require data to be passed through this separate buffer.

Instead of bothering with caching and reusing them, just allocate+free
them like stock and the FPC HAL.

Change-Id: I27f4cac1062a932dd820b5ec2bdf0b136b794f65
Signed-off-by: MarijnS95 <marijns95@gmail.com>