Egistec: Nile: Extract thread state management into separate class.
Extract thread state management to be reused in the upcoming Ganges
Egistec implementation.

In the future, if we can get Egistec devices to support gestures (or
just to reduce code bloat and make everything consistent), the extra
locking and condition_variable sync from FPC should be moved over into
this file. Obviously this implies a refactoring on FPC to utilize the
same implementation.

Change-Id: I748f19c8e6b789aebe65f0b446b2211315cde94d
Signed-off-by: MarijnS95 <marijns95@gmail.com>