Egistec: Ganges: Add state cleanup, fix concurrency issues.
Cleanup that had to be amended to the PR. We don't want to send
SetWorkMode(2) before all setup is done, thus postpone the thread start
to after this.
State cleanup (uninitializing everything) is mostly added for posterity,
may we ever need it.

Also fix the Enable() call at the beginning: This was an
Enable()+Reset() before, but the Reset() (instead of the Enable()) was
incorrectly replaced by a DeviceEnableGuard.

Change-Id: Ieb6e5abb88e76da562b09cfe0c21d4d5e7a67f88
Signed-off-by: MarijnS95 <marijns95@gmail.com>