BiometricsFingerprint: Never consume state when checking for cancel.
When reading an eventfd, the value is atomically reset to zero. In case
something different than STATE_CANCEL is requested, this state is
completely ignored.
Instead, only check whether an event is available, as that automatically
means something else has to happen. The relevant loops break out and
return control back to the worker thread which is responsible for
dealing with it consistently and appropriately.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>