WorkerThread: Replace IsCanceled with peeking IsEventAvailable.
Do not consume events to check if it is canceled. While no other events
are expected in the current implementation, as the (removed) log line
states it will accidentally consume any other event for no good reason.
Instead, poll the eventfd to figure out whether another event is
requested. If there is, this implies a cancellation of the current
state, followed by the thread handler consuming the state and acting
appropriately.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>