egistec: nile: New stock firmware now uses the Ganges HAL.
The TrustZone application that comes from the firmware partition is
updated by (one of) the latest stock updates, switching it over to the
same "new" API that's also in use by Ganges.
Check for this by "reading" a random number. If the size field (which
probably means something entirely different) is not set to the expected
value of 8 bytes, this must be the other HAL.

Note: There are perhaps other ways to detect this, but this is the
earliest error that is thrown, and seemingly the easiest and/or most
harmless (ExtraCommands have 0xa in the first byte, whereas all Ganges
trustlet_buffer_t calls start with 0xe0).

Signed-off-by: MarijnS95 <marijns95@gmail.com>