ion_buffer: Implement with backwards-compatible libion.
libion is a kernel-agnostic library that provides helpers for
allocation, deallocation, mapping and more. Instead of relying on kernel
headers to define TARGET_ION_ABI_VERSION, it checks the presence of a
new/legacy interface by testing the deprecated ION_IOC_FREE ioctl (the
only way to make a kernel-agnostic library is by checking at runtime).
Using this library is the way forward if we want to maintain an
implementation compatible with Linux 4.9, 4.14 and whatever the future
holds.

One of the major changes in the new API is the removal of the user
handle. Instead all buffers are referred by their fd. This provides
significant improvements to the API since an fd can be:
- used as handle;
- shared between processes;
- mmap()'ed;
- close()'d.
It greatly decreases the number of ioctls related to buffers (merely one
for allocation, and an unrelated one for heap queries) since the other
functionalities are either inferred or available through standard unix
calls.

Unfortunately ion_map() rejects any call on a non-legacy (4.12+) kernel,
instead of doing just the mmap. This is most likely because it passes
back _a separate fd_ acquired through ION_IOC_MAP which needs to be
closed separately (and does not exist in the non-legacy implementation).
Instead of adding a switch in the code that manages the extra fd on
legacy kernel builds, use ion_alloc_fd(). It "shares" the allocated user
handle on legacy kernels, giving a sharable fd that can also be mapped.
Curiously, it is possible to immediately close the ion user handle after
sharing, since that keeps the buffer open.
On non-legacy kernel implementations it simply calls the new allocation
ioctl, which returns a mappable fd immediately.

My only gripe with libion is that it calls its helper function ion_free
(which in turn uses ion_ioctl) to test for a legacy implementation. This
function logs a nasty, _useless_ error message:
E ion     : ioctl c0044901 failed with code -1: Invalid argument

Signed-off-by: MarijnS95 <marijns95@gmail.com>